Cleaning a bit

+++ build.sh
@@ -9,8 +9,11 @@ then
   if [ -e $FILENAME ];
   then
     docker build -f $FILENAME -t $REPONAME:$NAME .
-    VERSION=$(grep -oP 'Using Qt version \K[0-9.]+' <<< $(docker run --rm $REPONAME:$NAME qmake --version))
-    docker image tag $REPONAME:$NAME $REPONAME:$(sed "s/latest/$VERSION/" <<< $NAME)
+    if [ $NAME != "base" ]
+    then
+      VERSION=$(grep -oP 'Using Qt version \K[0-9.]+' <<< $(docker run --rm $REPONAME:$NAME qmake --version))
+      docker image tag $REPONAME:$NAME $REPONAME:$(sed "s/latest/$VERSION/" <<< $NAME)
+    fi
   else
     echo "No such Dockerfile found!";
   fi
